refactor(OmokRule): 오목 룰 라이브러리 복원
